requirements.txt
Pillow==3.0.0
falcon==0.3.0
gunicorn==19.3.0
python-mimeparse==0.1.4
selenium==2.48.0
